+++ include/CameraMetadata.h
@@ -68,7 +68,7 @@ struct ExtraData {
 
 struct CameraConfiguration {
     std::vector<float> apertures;
-    std::array<unsigned short, 4> blackLevel;
+    std::array<float, 4> blackLevel;
     std::array<float, 9> calibrationMatrix1;
     std::array<float, 9> calibrationMatrix2;
     std::string colorIlluminant1;
@@ -82,7 +82,7 @@ struct CameraConfiguration {
     std::array<float, 9> forwardMatrix2;
     int numSegments;
     std::string sensorArrangement;
-    unsigned short whiteLevel;
+    float whiteLevel;
 
     static CameraConfiguration parse(const std::string& jsonString);
     static CameraConfiguration parse(const nlohmann::json& j);
